Stark Refresh
After the TAFE assignment was a success, i have decided to scrap this
version of Stark as i have lernt alot more from this failed atempt and
have alot i would like to change.

I will possibly be making a Stark - JME3 Prototype in the near future,
before comitting to a full C++ version of Stark.

In the meantime, This has been emptied and will be left prepared for
what is to come.